[BUGFIX] Recursively expire pathcache entries
If a page is either moved or modified in a way that RealURL feels it has
an influence on its path generation (according to segTitleFieldList), an
expiration of the entry in tx_realurl_patchcache is triggered.

If a page with child pages is moved/modified, the expiration is only
done for the moved/modified page itself. This leads to a problem,
when a branch is moved and a branch with the exact same structure is
created again.

Like

Current Week
+ Top stories
+ Other stuff
Archive
+ ...

gets moved/renamed to another page and the structure

Current Week ( <- new created )
+ Top stories ( <- new created )
+ Other stuff ( <- new created )
Archive
+ Week 12 ( <- was Current Week before )
  + Top stories
  + Other stuff

The links to the sub-pages of the new page will look correct, but lead
to the old pages, as they haven't been expired, yet.

Change-Id: Id40a2f1a18e57889d53082b126f2ed9e6b93321f
Fixes: #41477
Reviewed-on: https://review.typo3.org/15177
Reviewed-by: Dmitry Dulepov
Tested-by: Dmitry Dulepov