[BUGFIX] Revitalization of pathcache confused if nothing changed
If a path expires because of content changes that don't
affect its URL, the old URL is marked as "expired" but
the new one is not inserted.
This happens e.g. if a "speaking url path segment" is
given for a page and the page title property is changed.
Then the URL is marked as "expired" because "title" might
affect the URL, but the new URL equals the old one because
"title" isn't used but "speaking url path segment".

Change-Id: Id58a56460e6a12b24d837fa4a3487fb9428d1989
Fixes: #42295