[TASK] Pass setup array to userFunc when processing path parts
This change is non-destructive; it simply adds one more parameter
value when calling user functions from the decode/encode method.

The reason for this change is the wish to pass arguments to the
user function being called in addition to the value and parent
object references, which sometimes are insufficient to affect
path part translation.

Change-Id: Ie3f101280d5d4d36a91af8a531352bca8ab70c2f
Fixes: #42496
Reviewed-on: https://review.typo3.org/16017
Reviewed-by: Dmitry Dulepov
Tested-by: Dmitry Dulepov