[BUGFIX] mysqli compatibility issue with TYPO3 CMS 6.1
As of TYPO3 CMS, the $GLOBALS['TYPO3_DB']->link is protected
(and not a resource anymore due to the switch to mysqli).

Thus, the check in realurl, if the extended mysql syntax should
be used (= enabled when using MySQL 4.1.0 or later and no DBAL),
the check should be different.

As realurl only supports TYPO3 4.5 and later, and 4.5 requires
mysql 5.0, the check is not needed anymore like it is.

Change-Id: I10eda5c206fe0a14c901a560359bd66a48330179
Fixes: #46807
Reviewed-on: https://review.typo3.org/19452
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack
Reviewed-by: Oliver Hader
Reviewed-by: Stefan Neufeind
Reviewed-by: Wouter Wolters
Tested-by: Oliver Hader
Reviewed-by: Tilo Baller
Tested-by: Tilo Baller
Reviewed-by: Ingo Pfennigstorf
Tested-by: Ingo Pfennigstorf
Reviewed-by: Dmitry Dulepov
Tested-by: Dmitry Dulepov