urldecode path segments separately
On decoding, the full scriptpath is handeld by urldecode(). After this,
the slash '/' is used as separator. Doing this way, realurl cannot
handle proper escaped slashes in values.

This patch does the decoding on every value separately as it did realUrl
1.x.

Imagin this scenario:

http://example.com/doc/http%3A%2F%2Fanotherexample.com%2Fapi%2Fdoc%3A123%2F6%2F/

On page doc, the following fixed post vars are configures:

'fixedPostVars' = array(
	'presentationConfiguration' => array(
                'docid' => array(
                        'GETvar' => 'tx_example[docid]',
                        ),
        ),
	20 => 'presentationConfiguration',
),

The extensions needs to get the passed URL in variable
tx_example[docid].