[TASK] In case of duplicated urldata entries, prefer those with cHash
It could happen that RealUrl encodes more than one urldata entry for the same speaking url. Usually one of the entries does have cHash and the other does not. This commit makes sure that, in this case, the entry with the cHash is taken into account over the one without.