Fix replacing wrong URL parts
Very important addition: Search for the bad URL part starting from the back instead of from the front of the string. This lead to wrong URLs if the bad URL part was e.g. a single character which is also part of the valid base URL:
If e.g. the valid URL is `http://example.org/en/imprint/index.html` and `http://example.org/en/imprint/in/index.html` is requested, without the patch the bad part `in` is not replaced where it is actually wrong but in the word `imprint`, so you got redirected to `http://example.org/en/impt/in/index.html`. By starting to search from the back the correct bad part is found and you get redirected to `http://example.org/en/imprint/index.html`