Set hostnames after automatic configuration was performed
Setting the hostnames before the automatic configuration was performed can cause exceptions. In the case when the realurl_autoconf.php is missing and the hostname ist still NULL then an exception 'RealURL was not able to find the root page id for the domain "www.example.com"' is thrown. The reason for this is that only after the automatic configuration was performed the hostname can be determined.