[BUGFIX] fix missing cHash configuration issue
If the TYPO3 bootstrap mechanism wasn't executet completely yet,
the check for gclid in cacheHash->excludedParameters ends
up in an php error, because the array isn't generated yet.
This happens for example with helhum/typo3_console help
command.

To solve this, add a condition to check if the array even exists.