[FEATURE] Alias per page
With this commit the page id is stored with each alias. So an alias can be
used more than once as it is identified by the page on which the record is shown.
With this commit the behaviour of aliases is similar to the beaviour of pages:
The path segment is unique inside of the parent page.